requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
yt-dlp==2024.12.23
python-multipart==0.0.18
websockets==14.1
aiofiles==24.1.0
pydantic==2.10.4
httpx==0.27.2
pycryptodome==3.21.0
